[DEV] Fix usable range in string-octoal size

+++ lib/util.c
@@ -163,7 +163,7 @@ oct_to_size(char *oct)
 
 /* integer to string-octal conversion, no NULL */
 void
-int_to_oct_nonull(int num, char *oct, size_t octlen)
+int_to_oct_nonull(long num, char *oct, size_t octlen)
 {
 	snprintf(oct, octlen, "%*lo", (int)(octlen - 1), (unsigned long)num);
 	oct[octlen - 1] = ' ';
